requirements.txt
# Core Dependencies
fastapi==0.104.1
uvicorn[standard]==0.24.0
pydantic==2.5.0
pydantic-settings==2.1.0
python-multipart==0.0.6

# Streamlit UI
streamlit==1.29.0

# LLM and Embeddings
langchain==0.1.0
langchain-community==0.0.10
sentence-transformers==2.2.2
transformers==4.36.0
torch==2.1.1

# Vector Database
chromadb==0.4.18

# Document Processing
pypdf==3.17.1
pypdf2==3.0.1
beautifulsoup4==4.12.2
lxml==4.9.3
python-docx==1.1.0
markdown==3.5.1
unstructured==0.11.4

# Web Scraping
requests==2.31.0
aiohttp==3.9.1
playwright==1.40.0

# Evaluation
ragas==0.1.1
langsmith==0.0.69
datasets==2.15.0

# Utilities
python-dotenv==1.0.0
pyyaml==6.0.1
tenacity==8.2.3
tqdm==4.66.1

# Observability
opentelemetry-api==1.21.0
opentelemetry-sdk==1.21.0
opentelemetry-instrumentation-fastapi==0.42b0
prometheus-client==0.19.0

# Guardrails
langkit==0.0.23
presidio-analyzer==2.2.33
presidio-anonymizer==2.2.33

# Testing
pytest==7.4.3
pytest-asyncio==0.21.1
httpx==0.25.2

